backend/requirements.txt
flask>=3.0
flask-cors>=4.0
google-generativeai>=0.8
google-genai
python-dotenv>=1.0
redis>=5.0
googlemaps>=4.10.0
PyJWT>=2.8.0
cachetools>=5.3.0
